% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Ajacency Matrix}
\usage{
adjacency_matrix(bigrams, binary = TRUE, left = FALSE)
}
\arguments{
\item{bigrams}{Bigram dataframe form ocunR::bigrams()}

\item{binary}{TRUE if binary, FALSE if frequency of bigram}

\item{left}{TRUE if left direction, FALSE if right direction}
}
\value{
A matrix with the names of each morph/word as row/column names.
}
\description{
Creates an adjacency matrix from a bigram dataframe
(see \code{\link{bigrams}}). The matrix's rows are the nodes and the columns
are the neighbors.
}
