% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigram.R
\name{bigrams}
\alias{bigrams}
\title{Dataframe of Bigrams}
\usage{
bigrams(ocun.data, morphs = TRUE, sentence.separators = FALSE)
}
\arguments{
\item{ocun.data}{Ocun dataframe column}

\item{morphs}{TRUE if considering unigrams as morphs, FALSE if considering unigrams as words}

\item{sentence.separators}{FALSE if not to consider sentence breaks, TRUE if otherwise}
}
\description{
Requires a column from an Ocun Data dataframe, normally generated with ocunR::read_txt.
}
