% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_chunker.R
\name{dm_chunker}
\alias{dm_chunker}
\title{De Marcken style chunker}
\usage{
dm_chunker(text, zipfian = FALSE, rate = 25, sep = "[.,;?!\\n]", iter = 10)
}
\description{
Splits a string into substrings based on MDL
}
