% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksm.R
\name{ksm}
\alias{ksm}
\title{Kernel Similarity Matrix (KSM)}
\usage{
ksm(adj.mat)
}
\arguments{
\item{adj.mat}{the adjacency matrix (see \link{adjacency_matrix}).}
}
\description{
Builds a kernel similarity matrix fom adjacency matrix
}
