% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon.daaichin}
\alias{lexicon.daaichin}
\title{Daai Chin Lexicon}
\format{
A named numeric vector with the costs of each word (plog).
}
\usage{
lexicon.daaichin
}
\description{
From Daai Chin descriptive grammar. Costs estimated by fitting a zipf distribution.
}
\keyword{datasets}
