% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon.english}
\alias{lexicon.english}
\title{English Lexicon}
\format{
A large numeric vector with the costs of each word. Words themselves are element names.
}
\source{
\url{http://tinypaste.com/c1666a6b}
}
\usage{
lexicon.english
}
\description{
From Generic Human's wordlist. costs estimated by fitting a zipf distribution.
}
\keyword{datasets}
