% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexicon}
\alias{lexicon}
\title{Lexicon}
\usage{
lexicon(tokens, zipf = TRUE)
}
\arguments{
\item{tokens}{Vector of tokens}

\item{zipf}{if TRUE will obtain token cost from zipf distribution.}
}
\description{
Creates a lexicon for a vector of tokens, especially useful for \link{parse}
}
