% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchunker_parser.R
\name{parse}
\alias{parse}
\title{Parse}
\usage{
parse(lexicon, str)
}
\arguments{
\item{lexicon}{The lexicon, see \link{lexicon}}

\item{str}{The string to be parsed}
}
\description{
Parse the lexical items in a string without spaces using a lexicon. NOTE:
this is a R implementation of the algorithm suggested by Generic Human
at StackExchange for Python. Kudos to him! See: \url{https://stackoverflow.com/questions/8870261/how-to-split-text-without-spaces-into-list-of-words/11642687#11642687}
}
