% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_txt.R
\name{read_txt}
\alias{read_txt}
\title{Read a Text File with Linguistic Data}
\usage{
read_txt(txt.file)
}
\arguments{
\item{txt.file}{Text file containing the data.}
}
\value{
A data frame with three columns: language, gloss and translation
}
\description{
File must comply with the following format: 3 lines per data, first line is the sentence in the
language with morphological segmentation, second line is the gloss and third line is the translation.
This function will therefore reject files with a not divisible by three count of lines.
}
