% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigram.R
\name{words}
\alias{words}
\title{Vector of Words}
\usage{
words(ocun.data, sentence.separators = FALSE)
}
\arguments{
\item{ocun.data}{Ocun dataframe column}

\item{sentence.separators}{FALSE if not to consider sentence breaks, TRUE if otherwise}
}
\description{
Requires a column from an Ocun Data dataframe, normally generated with ocunR::read_txt.
}
